DRY magic for social links
